+++ CHANGELOG.md
@@ -1,2 +1,4 @@
 # What's new:
 - Improved the speed of chunk meshing when their are block entities
+- Improved settings behaviour
+- Fixed issues on play.modernbeta.org